Backward selection in decomposable graphical models

Backward selection in decomposable graphical models

bws(df, adj, p = 0.5, ht = new.env(hash = TRUE), trace = TRUE,
  thres = 5)

Arguments

df
data.frame

adj
An adjacency list

p
Penalty term in the stopping criterion (0 = AIC and 1 = BIC)

ht
An environment (hashtable) containing precomputed entropies

trace
Logical indicating whether or not to trace the procedure

thres
A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.

Value

A bws object (a list) with values:

G_adj

The fitted graph as an adjacency list.

G_A

The fitted graph as an adjacency matrix.

e

The edge that was just deleted and the entropy attached to it.

S

The minimal separator that now separates the nodes in e.

C

A list with the (maximal) cliques of the graph.

ht

An updated version of the input argument ht.

Details

See efs for details about thres.

See also

Examples

d <- tgp_dat[, 5:8]
bws(d, make_complete_graph(colnames(d)))

#>  Edges: 5 of 6 - xic = -19.17848 

#>  A Decomposable Graph With 
#>  ------------------------- 
#>   Nodes: 4 
#>   Edges: 4 / 6 
#>   Cliques: 2 
#>   <efs> 
#>  -------------------------