Simulate observations from a decomposable graphical model

This function simulates observations from a DGM

dgm_sim(A, adj, nsim = 1000, ncores = 1)

Arguments

A
Character Matrix (data)

adj
Adjacency list of a decomposable graph

nsim
Number of simulations

ncores
Number of cores to use in parallelization

Value

This function returns a matrix of dimension nsim x ncol(A) where each row correspond to a simulated observation from a DGM represented by adj.

Examples

G  <- adj_list(efs(tgp_dat[, 5:8]))

#>  Edges: 2 of 6 - xic = -205.695545 
#>  Edges: 3 of 6 - xic = -111.017746 
#>  Edges: 4 of 6 - xic = -20.626792 
#>  Edges: 5 of 6 - xic = -0.970214 

dgm_sim(as.matrix(tgp_dat[, 5:8]), G, nsim = 10)

#>       rs260694 rs11123719 rs11691107 rs12469721
#>  [1,] "T"      "T"        "C"        "T"       
#>  [2,] "T"      "T"        "C"        "T"       
#>  [3,] "T"      "T"        "C"        "A"       
#>  [4,] "T"      "T"        "C"        "A"       
#>  [5,] "T"      "T"        "C"        "A"       
#>  [6,] "T"      "T"        "C"        "A"       
#>  [7,] "T"      "T"        "C"        "A"       
#>  [8,] "T"      "T"        "C"        "A"       
#>  [9,] "T"      "T"        "C"        "A"       
#> [10,] "T"      "T"        "C"        "A"