Efficient forward selection in decomposable graphical models

Efficient forward selection in decomposable graphical models

efs(df, x = efs_init(df), p = 0.5, trace = TRUE, thres = 5)

Arguments

df
data.frame

x
A efs object

p
Penalty term in the stopping criterion (0 = AIC and 1 = BIC)

trace
Logical indicating whether or not to trace the procedure

thres
A threshold mechanism for choosing between two different ways of calculating the entropy.

Value

A efs object (a list) with values:

G_adj

The fitted graph. Can be extracted with as_adj_lst

G_A

The edge that was just deleted and the entropy attached to it. Can be extracted using as_adj_mat

CG

The cliques of the graph.

CG_A

A neighbor matrix of the clique graph.

MSI

Minimal separator info list

ht

An updated version of the input argument ht.

Details

The algorithm relies on calculating lots of entropies (see https://arxiv.org/abs/1301.2267). If thres is equal to a number, say a, the entropy will be calculated using the table function for all situations where the number of variables is less than or equal to a. Otherwise, the entropies will be calculated by pasting rows in the subset of df over the relevant variables. For small values of a the former method is generally faster whereas for large values of a it breaks down and it is necessary to use the latter method.

References

https://arxiv.org/abs/1301.2267, https://doi.org/10.1109/ictai.2004.100

See also

Examples

efs(tgp_dat[, 5:8])

#>  Edges: 1 of 6 - xic = -205.695545 
#>  Edges: 2 of 6 - xic = -111.017746 
#>  Edges: 3 of 6 - xic = -20.626792 
#>  Edges: 4 of 6 - xic = -0.970214 

#>  A Decomposable Graph With 
#>  ------------------------- 
#>   Nodes: 4 
#>   Edges: 5 / 6 
#>   Cliques: 2 
#>   <efs> 
#>  -------------------------