Fit a decomposable graphical model

A generic method for structure learning in decomposable graphical models

fit_graph(
  df,
  type = "fwd",
  adj = NULL,
  q = 0.5,
  trace = TRUE,
  thres = 5,
  wrap = TRUE
)

Arguments

df
data.frame

type
Character ("fwd", "bwd", "tree" or "tfwd")

adj
Adjacency list of a decomposable graph

q
Penalty term in the stopping criterion (0 = AIC and 1 = BIC)

trace
Logical indicating whether or not to trace the procedure

thres
A threshold mechanism for choosing between two different ways of calculating the entropy.

wrap
logical specifying if the result of a run with type = "tree" should be converted to a "fwd" object

Value

A gengraph object

Details

The types are

"fwd": forward selection

"bwd": backward selection

"tree": Chow-Liu tree (first order interactions only)

"tfwd": A combination of "tree" and "fwd". This can speed up runtime considerably in high dimensions.

Using adj_lst on an object returned by fit_graph gives the adjacency list corresponding to the graph. Similarly one can use adj_mat to obtain an adjacency matrix. Applying the rip function on an adjacency list returns the cliques and separators of the graph.

References

https://arxiv.org/abs/1301.2267, https://doi.org/10.1109/ictai.2004.100

See also

Examples

# NOT RUN {
library(dplyr)

## All handwritten digits that have true class equal to a "1".
d <- digits %>%
  filter(class == "1") %>%
  select(-class)

g <- fit_graph(d)
print(g)
plot(g, vertex.size = 1)
adjm <- adj_mat(g)
adjl <- adj_lst(g)

## Cliques and separators in the graph
rip(adjl)
# }