Outlier detection

Detecting outliers within a dataset or test if a new (novel) observation is an outlier.

fit_outlier(
  A,
  adj,
  z = NULL,
  alpha = 0.05,
  nsim = 10000,
  ncores = 1,
  validate = TRUE
)

Arguments

A
Character matrix or data.frame. All values must be limited to a single character.

adj
Adjacency list or gengraph object of a decomposable graph. See package ess for gengraph objects.

z
Named vector (same names as colnames(A)) or NULL. See details. Values must be limited to a single character.

alpha
Significance level

nsim
Number of simulations

ncores
Number of cores to use in parallelization

validate
Logical. If true, it checks if A only has single character values and converts it if not.

Details

If the goal is to detect outliers within A set z to NULL; this procedure is most often just referred to as outlier detection. Once fit_outlier has been called in this situation, one can exploit the outliers function to get the indicies for which observations in A that are outliers. See the examples.

On the other hand, if the goal is test if the new unseen observation z is an outlier inA, then supply a named vector to z.

All values must be limited to a single character representation; if not, the function will internally convert to one such representation. The reason for this, is a speedup in runtime performance. One can also use the exported function to_single_chars on A in advance and set validate to FALSE.

The adj object is most typically found using fit_graph from the ess package. But the user can supply an adjacency list, just a named list, of their own choice if needed.

See also

Examples