Probability mass function for a decomposable graphical model

Probability mass function for a decomposable graphical model

pmf(A, adj, logp = FALSE)

Arguments

A
A character matrix of data

adj
Adjacency list or gengraph object of a decomposable graph. See fit_graph.

logp
Logical; if TRUE, probabilities p are given as log(p).

Value

A function - the probability mass function corresponding to the decomposable graph adj using A to estimate the probabilities.

Details

It should be noted, that pmf is a closure function; i.e. its return value is a function. Once pmf is called, one can query probabilities from the returned function. If the probability of an observation is zero, the function will return NA if logp is set to TRUE and zero otherwise.

Examples

# NOT RUN {
library(dplyr)

# All handwritten digits that have true class equal to a "1".
d <- digits %>% # subset(digits, class == "1")
  filter(class == "1") %>%
  select(-class)

# A handwritten digit with true class equal to "1"
z1 <- digits %>%
  filter(class == "1") %>%
  select(-class) %>%
  slice(5) %>%
  unlist()

# A handwritten digit with true class equal to "7"
z7 <- digits %>%
  filter(class == "7") %>%
  select(-class) %>%
  slice(1) %>%
  unlist()

# Fit an interaction graph
g <- fit_graph(d, trace = FALSE)
plot(g)

g <- g %>%
  adj_lst()

# Probability in class "1"
p1 <- pmf(d %>% as.matrix(), g)
p1(z7)
p1(z1)

# Probability on component 23 in class "1"
cmp   <- components(g)
cmp23 <- cmp[[23]]
print(cmp23)
p1_23 <- pmf(d %>% select(cmp23) %>% as.matrix(), g[cmp23])
p1_23(z7)
p1_23(z1)

# }