Fit Outlier

The outlier model relies on a graph object returned from fit_graph. This function needs data; however when an outlier test is performed, the function fit_graph needs data with z appended. That is because, the model is under the hypothesis that z is not an outlier in A. fit_outlier automatically do this. Moreover, fit_outlier relies on outlier_model which is not an exported function, but can be used carefully by the user for more flexibillity (use molic:::outlier_model). The user should be aware, that it is assumed that the data argument in outlier_model has the new observation included! Thus fit_outlier is a safe wrapper for most applications.

fit_outlier(
  A,
  z,
  alpha = 0.05,
  type = "fwd",
  q = 0.5,
  comp = NULL,
  nsim = 10000,
  ncores = 1,
  trace = FALSE,
  validate = TRUE
)

Arguments

A
Character Matrix (data) - without the new observation z appended

z
Named vector (same names as colnames(A))

alpha
The significance level

type
Character ("fwd", "bwd", "tree" or "tfwd") - see fit_graph

q
Penalty term in the stopping criterion (0 = AIC and 1 = BIC) - see fit_graph

comp
A list with character vectors. Each elementer in the list is a component in the graph (using expert knowledge)

nsim
Number of simulations

ncores
Number of cores to use in parallelization

trace
Logical indicating whether or not to trace the procedure

validate
Logical. If true, it checks if A has only single character values and converts it if not.

Details

A convinient wrapper for doing outlier test

See also

Examples

# NOT RUN {
library(dplyr)

# All handwritten digits that have true class equal to a "1".
d <- digits %>%
  filter(class == "1") %>%
  select(-class)

# A handwritten digit with true class equal to "7"
z <- digits %>%
  filter(class == "7") %>%
  select(-class) %>%
  slice(1) %>%
  unlist()

# Test if z is an outlier in class "1"
m <- fit_outlier(as.matrix(d), z)
print(m)
pmf(m)
# }